refactor: support extra component
